+++ src/Helpers/Types/Dense_tensor.h
@@ -17,13 +17,13 @@ private:
   int type_id = -1; // Maybe, it can be changed to the dimension of the type
 
   /// Shape of the tensor
-  std::vector<int> shape;
+  std::vector<long> shape;
 
 public:
   /// Construct the tensor from the type id and the shape
   /// \param in_type_id onnx::TensorProto_DataType_* id
   /// \param in_shape Shape of the tensor
-  Dense_tensor(int in_type_id, std::vector<int> in_shape);
+  Dense_tensor(int in_type_id, std::vector<long> in_shape);
 
   /// Construct the tensor from a onnx::ValueInfoProto object
   /// \param info onnx::ValueInfoProto object
